refactor: WBS 컴포넌트를 wbs 폴더로 분리
- components/wbs/ 폴더 생성
- hierarchical-wbs-table.tsx, wbs-table.tsx 이동
- project-view.tsx의 import 경로 업데이트
